pyproject.toml
[tool.poetry]
name = "tanks2d"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
pygame = "2.5.2"
beautifulsoup4 = "4.12.2"
soupsieve = "2.5"
mysql-connector = "2.2.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
